pluto-backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pluto-backend"
version = "1.0.0"
description = "Pluto API - Multi-mode social platform"
requires-python = "==3.12.*"
dependencies = [
    # FastAPI core — [standard] installs fastapi-cli (the 'fastapi' binary)
    "fastapi[standard]>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database — asyncpg is the async driver; psycopg2-binary for sync fallback
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",

    # Firebase Admin
    "firebase-admin==6.5.0",

    # Validation & settings
    "pydantic==2.7.1",
    "pydantic-settings==2.3.0",
    "email-validator==2.1.1",

    # HTTP client
    "httpx==0.27.0",

    # Networking
    "websockets==12.0",

    # Auth
    "python-jose[cryptography]==3.3.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",

    # Supabase
    "supabase==2.5.0",

    # Google Cloud Storage
    "google-cloud-storage==2.16.0",

    # Security & rate limiting
    "slowapi==0.1.9",
    "python-dotenv==1.0.1",

    # Utilities
    "Pillow==10.3.0",
    "python-dateutil==2.9.0",
    "ujson==5.10.0",
    "structlog==24.1.0",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["app"]

# FastAPI Cloud entrypoint — app/main.py is auto-detected but explicit is safer
[tool.fastapi]
entrypoint = "app.main:app"